requirements.txt
numpy
pandas
openai
pinecone-client[grpc]
tqdm